Add ImageURLProcessor to replace actualPhoto for PlayerPreview
