+++ Betterpick/common/Networking/Manager/BetterpickAPIManagerMock.swift
@@ -105,7 +105,7 @@ class BetterpickAPIManagerMock: BetterpickAPIManager {
 
     // MARK: GET /players/{playerID}
     override func player(playerID: String, completion: @escaping BetterpickAPIManager.Callback<Player>) {
-        let player = Player(id: 0, name: "L. Messi", age: 35, photo: URL(string: "https://cdn.sofifa.com/players/158/023/20_120.png")!, nationality: "Argentina", overall: 94, club: "FC Barcelona", value: 95000000, wage: 540000, releaseClause: 9999, preferredFoot: "Left", skillMoves: 5, workRate: "Medium/Low", position: .CF, jerseyNumber: 10, height: "kek", weight: 9, crossing: 85, finishing: 98, headingAccuracy: 83, shortPassing: 90, volleys: 90, dribbling: 97, curve: 96, fkAccuracy: 95, longPassing: 90, ballControl: 98, acceleration: 93, sprintSpeed: 89, agility: 90, reactions: 91, balance: 90, shotPower: 88, jumping: 84, stamina: 80, strength: 85, longshots: 93, aggression: 84, interceptions: 85, positioning: 94, vision: 95, penalties: 94, composure: 98, marking: 93, standingTackle: 88, slidingTackle: 82, gkDiving: 46, gkHandling: 48, gkKicking: 60, gkPositioning: 44, gkReflexes: 41)
+        let player = Player(id: 0, name: "L. Messi", age: 35, photo: URL(string: "https://cdn.sofifa.org/players/4/19/158023.png")!, nationality: "Argentina", overall: 94, club: "FC Barcelona", value: 95000000, wage: 540000, releaseClause: 9999, preferredFoot: "Left", skillMoves: 5, workRate: "Medium/Low", position: .CF, jerseyNumber: 10, height: "kek", weight: 9, crossing: 85, finishing: 98, headingAccuracy: 83, shortPassing: 90, volleys: 90, dribbling: 97, curve: 96, fkAccuracy: 95, longPassing: 90, ballControl: 98, acceleration: 93, sprintSpeed: 89, agility: 90, reactions: 91, balance: 90, shotPower: 88, jumping: 84, stamina: 80, strength: 85, longshots: 93, aggression: 84, interceptions: 85, positioning: 94, vision: 95, penalties: 94, composure: 98, marking: 93, standingTackle: 88, slidingTackle: 82, gkDiving: 46, gkHandling: 48, gkKicking: 60, gkPositioning: 44, gkReflexes: 41)
         returnSuccessAfter(duration: 0.2, completion: completion, response: player)
     }
